sandbox v0.1.0 Ev

Very simple sync/async event sourcing library. The main purpose is to integrate some event sourcing into existing codebase with little to no everhead. It does not use any external event bus, everything happens in the same Beam VM with simple function calls (sync) or message passing (async).

You need two things to integrate this library:

Event definitions

Event handlers

Event definitions

Create a module (e.g. MyApp.Events), include use Ev passing you OTP app name (more on that later) and define you events. Under the hood each event is an Exliri struct with enforced keys and typespec.

defmodule MyApp.Events do
  use Ev, otp_app: :myapp

  # Define events with defev(Name, Attr: Type)

  defev UserSignedUp, user_id:    pos_integer,
                      user_name:  String.t

  defev OrderPlaced,  order_id: pos_integer,
                      user_id:  pos_integer,
                      value:    non_neg_integer
end

Event handlers

Event handlers are modules with a single handle(event) callback function. You can create as many event handlers as you want to.

defmodule MyApp.SlackNotifications do
  @behaviour Ev.Handler

  alias MyApp.Events

  def handle(%Events.UserSignedUp{} = ev) do
    IO.puts "SLACK: New user signed up! #{ev.user_name}"
  end

  def handle(%Events.OrderPlaced{} = ev) do
    IO.puts "SLACK: New order placed #{ev.value}"
  end

  # catch-all clause
  def handle(_), do: :ok
end

Configuration

To configure event handlers for different environments use standard mix project configuration. With this approach it is trivial to exclude sending emails or notifications in dev/test.

# config/dev.exs
config :myapp, MyApp.Events, handlers: [
  MyApp.Indexer
]

# config/prod.exs
config :myapp, MyApp.Events, handlers: [
  MyApp.Indexer,
  MyApp.SlackNotifications,
  MyApp.MarketingEmails
]

# config/test.exs
config :myapp, MyApp.Events, handlers: [
  MyApp.TestHandler
]

Publishing events

Use MyApp.Events.publish_sync/1 or publish_async/1 to

Functions

__using__(opts \\ []) (macro)

Define your events module

Example:

defmodule MyApp.Events do
  use Ev, otp_app: :myapp

  defev UserSignedUp, user_id: pos_integer, user_name: String.t
  defev OrderPlaced, order_id: pos_integer, value: non_neg_integer
end

defev(name, attrs \\ []) (macro)

Define module with struct and typespec, in single line

Example:

use Ev
defev User, id:   integer,
            name: String.t

is the same as

defmodule User do
  @type t :: %__MODULE__{
    id:   integer,
    name: String.t
  }
  @enforce_keys [:id, :name]
  defstruct [:id, :name]
end

publish_async(event, handlers)

Publish event to handlers asynchronously. Each handler will be run in a separate process.

publish_sync(event, handlers)

Publish event to handers synchronously