sandbox v0.1.0 Ok

Utility module to work with {:ok, value} | {:error, reason} values

Types

t()

t() :: {:ok, any} | {:error, any}

tenum()

tenum() :: {:ok, Enum.t} | {:error, any}

Functions

each(tuple, function)

each(t, (any -> any)) :: t

Run function if tuple is OK. Only useful for side effects

Examples

iex> Ok.each({:ok, 2}, fn x -> x * 5 end)
{:ok, 2}

iex> Ok.each({:error, "oups"}, fn x -> x * 5 end)
{:error, "oups"}

iex> {:ok, 123} |> Ok.map(&IO.inspect/1)
{:ok, 123}

error(reason)

error(any) :: {:error, any}

Wrap error into {:error, error} tuple

Examples

iex> Ok.error(:econnrefused)
{:error, :econnrefused}

flat_map(typle, function)

flat_map(t, (any -> t)) :: t

Flat map tuple with given fun only if it’s OK

Examples

iex> Ok.flat_map({:ok, 2}, fn x -> {:ok, x + 1} end)
{:ok, 3}

iex> Ok.flat_map({:ok, 2}, fn _ -> {:error, "oups"} end)
{:error, "oups"}

iex> Ok.flat_map({:error, "oups"}, fn x -> {:ok, x + 1} end)
{:error, "oups"}

flat_map_enum(tuple_with_enum, function)

flat_map_enum(tenum, (any -> any)) :: tenum

Flat map tuple with enum given fun only if it’s OK

Examples

iex> Ok.flat_map_enum({:ok, [1,2,3]}, fn x -> {:ok, x + 1} end)
{:ok, [2,3,4]}

iex> Ok.map_enum({:error, "oups"}, fn x -> x + 1 end)
{:error, "oups"}

iex> Ok.flat_map_enum {:ok, [1,2,3]}, fn
...>   1 -> {:ok, 1}
...>   x -> {:error, "#{x} is bad"}
...> end
{:error, ["2 is bad", "3 is bad"]}

flatten(nested_tuple)

flatten({:ok, t} | {:error, any}) :: t

Flatten (also called “join”) nested tuples into single one

Examples

iex> Ok.flatten({:ok, {:ok, 1}})
{:ok, 1}

iex> Ok.flatten({:error, {:ok, "hm"}})
{:error, {:ok, "hm"}}

iex> Ok.flatten({:ok, {:error, "oups"}})
{:error, "oups"}

get(tuple, fallback)

get(t, any) :: any

Get value or use the fallback

Exapmles

iex> Ok.get({:ok, 1}, 0)
1

iex> Ok.get({:error, :notfound}, 0)
0

map(tuple, function)

map(t, (any -> any)) :: t

Map tuple with given fun only if it’s OK

Examples

iex> Ok.map({:ok, 2}, fn x -> x * 5 end)
{:ok, 10}

iex> Ok.map({:error, "oups"}, fn x -> x * 5 end)
{:error, "oups"}

iex> {:ok, 123} |> Ok.map(& &1 + 321)
{:ok, 444}

map_enum(tuple_with_enum, function)

map_enum(tenum, (any -> any)) :: tenum

Map tuple with enum given fun only if it’s OK

Examples

iex> Ok.map_enum({:ok, [1,2,3]}, fn x -> x + 1 end)
{:ok, [2,3,4]}

iex> Ok.map_enum({:error, "oups"}, fn x -> x + 1 end)
{:error, "oups"}

ok(term)

ok(any) :: {:ok, any}

Wrap values with {:ok, value} tuple

Examples

iex> Ok.ok(1)
{:ok, 1}

seq(list_of_tuples)

seq([t]) :: {:ok, list} | {:error, any}

Convert list of ok/error tuples into ok/error tuple of list

Examples

iex> Ok.seq([{:ok, 1}, {:ok, 2}, {:ok, 3}])
{:ok, [1,2,3]}

iex> Ok.seq([{:error, "oups"}, {:error, "meh"}])
{:error, ["oups", "meh"]}

iex> Ok.seq([{:ok, 1}, {:error, "bad 2"}, {:error, "bad 3"}, {:ok, 4}])
{:error, ["bad 2", "bad 3"]}

iex> Ok.seq([])
{:ok, []}

tand(tuple, tuple)

tand(t, t) :: t

Boolean AND on tuples (tuple-and, tand)

Examples

iex> Ok.tand({:ok, 1}, {:ok, 2})
{:ok, 2}

iex> Ok.tand({:error, "oups"}, {:ok, 2})
{:error, "oups"}

iex> Ok.tand({:ok, 1}, {:error, "meh"})
{:error, "meh"}

iex> Ok.tand({:error, "oups"}, {:error, "meh"})
{:error, "oups"}

tor(tuple, tuple)

tor(t, t) :: t

Boolean OR on tuples (tuple-or, tor)

Examples

iex> Ok.tor({:ok, 1}, {:ok, 2})
{:ok, 1}

iex> Ok.tor({:error, "oups"}, {:ok, 2})
{:ok, 2}

iex> Ok.tor({:ok, 1}, {:error, "meh"})
{:ok, 1}

iex> Ok.tor({:error, "oups"}, {:error, "meh"})
{:error, "meh"}

wrap(value, reason)

wrap(any, any) :: t

Wrap value in {:ok, value} tuple. If the value is nil return {:error, reason} tuple instead

Exapmles

iex> Ok.wrap(2, :notfound)
{:ok, 2}

iex> Ok.wrap(nil, :notfound)
{:error, :notfound}