Sandbox.Accounts (sandbox v0.1.0)

Context for accounts

Functions

detail(conn, id)

@spec detail(map(), binary()) ::
  {:ok,
   %Sandbox.Accounts.AccountDetail{
     account_id: term(),
     account_number: term(),
     links: term(),
     routing_numbers: term()
   }}
  | :error

Account detail.

get_account(conn, id)

@spec get_account(map(), binary()) ::
  {:ok,
   %Sandbox.Accounts.Account{
     currency: term(),
     enrollment_id: term(),
     id: term(),
     institution: term(),
     last_four: term(),
     links: term(),
     name: term(),
     subtype: term(),
     type: term()
   }}
  | :error

Returns one account.

get_accounts_from_token(token)

@spec get_accounts_from_token(binary()) :: list()

Transforms the token into list of accounts ids.

list_accounts(conn)

@spec list_accounts(map()) :: [
  %Sandbox.Accounts.Account{
    currency: term(),
    enrollment_id: term(),
    id: term(),
    institution: term(),
    last_four: term(),
    links: term(),
    name: term(),
    subtype: term(),
    type: term()
  }
]

List accounts. Depending on the provided token between returns between 1 and 8 accounts.

It will return the same data if the same token is passed.

Here 2 kind of hashing algorithms are used.

:erlang.phash2() Very usefull for getting an integer from any string.

:crypto.hash(:md5, _) Used for calculating ID's.

transform_token_into_account(times, token)

@spec transform_token_into_account(number(), binary()) :: binary()

Hashes the token n times for different results